Readme file was created and filled with information about project and its configuration
