+++ src/main/java/mate/academy/internetshop/controller/Bucket/ShowBucketItemsController.java
@@ -1,4 +1,4 @@
-package mate.academy.internetshop.controller;
+package mate.academy.internetshop.controller.Bucket;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
@@ -6,30 +6,38 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-
 import mate.academy.internetshop.exception.DataProcessingException;
 import mate.academy.internetshop.lib.Inject;
+import mate.academy.internetshop.models.Bucket;
+import mate.academy.internetshop.models.User;
+import mate.academy.internetshop.services.BucketService;
 import mate.academy.internetshop.services.UserService;
 import org.apache.log4j.Logger;
 
-@WebServlet(urlPatterns = "/Servlet/deleteUser")
-public class DeleteUserController extends HttpServlet {
+@WebServlet(urlPatterns = "/Servlet/ShowBucketItems")
+public class ShowBucketItemsController extends HttpServlet {
     @Inject
     private static UserService userService;
-    private static final Logger LOGGER = Logger.getLogger(DeleteUserController.class);
+    @Inject
+    private static BucketService bucketService;
+    private static final Logger LOGGER = Logger.getLogger(ShowBucketItemsController.class);
 
     @Override
     protected void doGet(HttpServletRequest req,
                          HttpServletResponse resp)
             throws ServletException, IOException {
-        Long userId = Long.valueOf(req.getParameter("userId"));
+        Long userId = (Long) req.getSession().getAttribute("userId");
+        User user = null;
         try {
-            userService.delete(userId);
+            user = userService.get(userId);
         } catch (DataProcessingException e) {
             LOGGER.error(e);
             req.setAttribute("Msg", e);
             req.getRequestDispatcher("/WEB-INF/views/error.jsp").forward(req, resp);
         }
-        resp.sendRedirect(req.getContextPath() + "/Servlet/index");
+        Bucket bucket = bucketService.get(userId);
+        req.setAttribute("user", user);
+        req.setAttribute("bucket", bucket.getItems());
+        req.getRequestDispatcher("/WEB-INF/views/Bucket.jsp").forward(req, resp);
     }
 }
