+++ src/main/java/mate/academy/internetshop/controller/Items/CreateItemController.java
@@ -1,4 +1,4 @@
-package mate.academy.internetshop.controller;
+package mate.academy.internetshop.controller.Items;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
@@ -9,22 +9,25 @@ import javax.servlet.http.HttpServletResponse;
 
 import mate.academy.internetshop.exception.DataProcessingException;
 import mate.academy.internetshop.lib.Inject;
+import mate.academy.internetshop.models.Item;
 import mate.academy.internetshop.services.ItemService;
 import org.apache.log4j.Logger;
 
-@WebServlet(urlPatterns = "/Servlet/deleteItem")
-public class DeleteItemController extends HttpServlet {
+@WebServlet(urlPatterns = "/Servlet/createItem")
+public class CreateItemController extends HttpServlet {
     @Inject
     private static ItemService itemService;
-    private static final Logger LOGGER = Logger.getLogger(DeleteItemController.class);
+    private static final Logger LOGGER = Logger.getLogger(CreateItemController.class);
 
     @Override
-    protected void doGet(HttpServletRequest req,
-                         HttpServletResponse resp)
+    protected void doPost(HttpServletRequest req,
+                          HttpServletResponse resp)
             throws ServletException, IOException {
-        Long id = Long.valueOf(req.getParameter("itemId"));
+        String itemName = req.getParameter("itemName");
+        Double itemPrice = Double.valueOf(req.getParameter("itemPrice"));
+        Item newItem = new Item(itemName).setPrice(itemPrice);
         try {
-            itemService.delete(id);
+            itemService.create(newItem);
         } catch (DataProcessingException e) {
             LOGGER.error(e);
             req.setAttribute("Msg", e);
