+++ src/main/java/mate/academy/internetshop/controller/Order/CompleteOrderController.java
@@ -1,4 +1,4 @@
-package mate.academy.internetshop.controller;
+package mate.academy.internetshop.controller.Order;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
@@ -8,36 +8,36 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import mate.academy.internetshop.exception.DataProcessingException;
 import mate.academy.internetshop.lib.Inject;
-import mate.academy.internetshop.models.Bucket;
-import mate.academy.internetshop.models.User;
 import mate.academy.internetshop.services.BucketService;
+import mate.academy.internetshop.services.OrderService;
 import mate.academy.internetshop.services.UserService;
 import org.apache.log4j.Logger;
 
-@WebServlet(urlPatterns = "/Servlet/ShowBucketItems")
-public class ShowBucketItemsController extends HttpServlet {
+@WebServlet(urlPatterns = "/Servlet/CompleteOrder")
+public class CompleteOrderController extends HttpServlet {
+
     @Inject
-    private static UserService userService;
+    private static OrderService orderService;
     @Inject
     private static BucketService bucketService;
-    private static final Logger LOGGER = Logger.getLogger(ShowBucketItemsController.class);
+    @Inject
+    private static UserService userService;
+    private static final Logger LOGGER = Logger.getLogger(CompleteOrderController.class);
 
     @Override
     protected void doGet(HttpServletRequest req,
                          HttpServletResponse resp)
             throws ServletException, IOException {
-        Long userId = (Long) req.getSession().getAttribute("userId");
-        User user = null;
+        Long userId = (Long)req.getSession().getAttribute("userId");
         try {
-            user = userService.get(userId);
+            orderService.completeOrder(bucketService.get(userId).getItems(),
+                    userService.get(userId));
         } catch (DataProcessingException e) {
-            LOGGER.error(e);
             req.setAttribute("Msg", e);
+            LOGGER.error(e);
             req.getRequestDispatcher("/WEB-INF/views/error.jsp").forward(req, resp);
         }
-        Bucket bucket = bucketService.get(userId);
-        req.setAttribute("user", user);
-        req.setAttribute("bucket", bucket.getItems());
-        req.getRequestDispatcher("/WEB-INF/views/Bucket.jsp").forward(req, resp);
+        req.setAttribute("orders", orderService.getUserOrders(userId));
+        req.getRequestDispatcher("/WEB-INF/views/ShowAllOrders.jsp").forward(req, resp);
     }
 }
