+++ src/main/java/mate/academy/internetshop/controller/User/DeleteUserController.java
@@ -1,4 +1,4 @@
-package mate.academy.internetshop.controller;
+package mate.academy.internetshop.controller.User;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
@@ -9,25 +9,22 @@ import javax.servlet.http.HttpServletResponse;
 
 import mate.academy.internetshop.exception.DataProcessingException;
 import mate.academy.internetshop.lib.Inject;
-import mate.academy.internetshop.models.Item;
-import mate.academy.internetshop.services.ItemService;
+import mate.academy.internetshop.services.UserService;
 import org.apache.log4j.Logger;
 
-@WebServlet(urlPatterns = "/Servlet/createItem")
-public class CreateItemController extends HttpServlet {
+@WebServlet(urlPatterns = "/Servlet/deleteUser")
+public class DeleteUserController extends HttpServlet {
     @Inject
-    private static ItemService itemService;
-    private static final Logger LOGGER = Logger.getLogger(CreateItemController.class);
+    private static UserService userService;
+    private static final Logger LOGGER = Logger.getLogger(DeleteUserController.class);
 
     @Override
-    protected void doPost(HttpServletRequest req,
-                          HttpServletResponse resp)
+    protected void doGet(HttpServletRequest req,
+                         HttpServletResponse resp)
             throws ServletException, IOException {
-        String itemName = req.getParameter("itemName");
-        Double itemPrice = Double.valueOf(req.getParameter("itemPrice"));
-        Item newItem = new Item(itemName).setPrice(itemPrice);
+        Long userId = Long.valueOf(req.getParameter("userId"));
         try {
-            itemService.create(newItem);
+            userService.delete(userId);
         } catch (DataProcessingException e) {
             LOGGER.error(e);
             req.setAttribute("Msg", e);
