+++ src/main/java/mate/academy/internetshop/controller/User/ShowAllUsersController.java
@@ -1,4 +1,4 @@
-package mate.academy.internetshop.controller;
+package mate.academy.internetshop.controller.User;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
